Create a timeout class that uses the core ruby timeout code
This will modify that could though so the timer can be updated at any point in time and keep running. The ruby version only allowed you to input a fixnum. This is hard to do when executing an SSH task. So as long as the client is still getting input from SSH the timeout should not happen. This allows that to happen but still timeout when it is no longer reciving output.